Multithread tests moved to SymbolTableStoragingTest.kt
