+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/interpreter/SymbolTableStoragingTest.kt
@@ -3,6 +3,7 @@ package com.smeup.rpgparser.interpreter
 import com.smeup.rpgparser.execution.Configuration
 import com.smeup.rpgparser.execution.getProgram
 import org.junit.Test
+import java.util.concurrent.Executors
 import kotlin.test.Ignore
 import kotlin.test.assertEquals
 
@@ -347,4 +348,51 @@ class SymbolTableStoragingTest {
                 actual = variables["NUM"] ?: error("Not found NUM")
         )
     }
+
+    @Test
+    fun multiThreadTest() {
+
+        val testNames = listOf("execPgmAndEvaluateStorage","initPgmByStorageAndEvaluateResult",
+                "execLRPgmAndEvaluateStorage", "execRTPgmTwiceAndPreserveValues", "initPreExistingVariablesPgmByStorageAndEvaluateResult",
+                "sameVariablesButDifferentACTGRP", "sameVariablesAndSameACTGRP")
+
+        val fixedThreadPool = 10
+        val repeatTests = 100
+
+        val executor = Executors.newFixedThreadPool(fixedThreadPool)
+        repeat(repeatTests){
+            for (testName in testNames) {
+                var simbolTableStoragingTest = SymbolTableStoragingTest()
+                val workerThread: Runnable = WorkerThread(simbolTableStoragingTest, testName)
+                executor.execute(workerThread)
+            }
+        }
+        executor.shutdown()
+
+        println("Waiting for all thread to finish...")
+        while (!executor.isTerminated) {
+        }
+        println("...done")
+    }
+}
+
+class WorkerThread(var symbolTableStoragingTest: SymbolTableStoragingTest, var testName: String) : Runnable {
+
+    override fun run() {
+        println(Thread.currentThread().name + " Start test $testName")
+        when (testName) {
+            "execPgmAndEvaluateStorage" -> symbolTableStoragingTest.execPgmAndEvaluateStorage()
+            "initPgmByStorageAndEvaluateResult" -> symbolTableStoragingTest.initPgmByStorageAndEvaluateResult()
+            "execLRPgmAndEvaluateStorage" -> symbolTableStoragingTest.execLRPgmAndEvaluateStorage()
+            "execRTPgmTwiceAndPreserveValues" -> symbolTableStoragingTest.execRTPgmTwiceAndPreserveValues()
+            "initPreExistingVariablesPgmByStorageAndEvaluateResult" -> symbolTableStoragingTest.initPreExistingVariablesPgmByStorageAndEvaluateResult()
+            "sameVariablesButDifferentACTGRP" -> symbolTableStoragingTest.sameVariablesButDifferentACTGRP()
+            "sameVariablesAndSameACTGRP" -> symbolTableStoragingTest.sameVariablesAndSameACTGRP()
+            else -> {
+                print("Test $testName not exists" )
+            }
+        }
+        println(Thread.currentThread().name + " End test $testName")
+    }
+
 }
